product-api - done

+++ product-api/src/main/java/com/github/fabiomqs/productapi/modules/product/service/ProductService.java
@@ -1,19 +1,27 @@
 package com.github.fabiomqs.productapi.modules.product.service;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.fabiomqs.productapi.config.exception.SuccessResponse;
 import com.github.fabiomqs.productapi.config.exception.ValidationException;
 import com.github.fabiomqs.productapi.modules.category.service.CategoryService;
-import com.github.fabiomqs.productapi.modules.product.dto.ProductRequest;
-import com.github.fabiomqs.productapi.modules.product.dto.ProductResponse;
+import com.github.fabiomqs.productapi.modules.product.dto.*;
 import com.github.fabiomqs.productapi.modules.product.model.Product;
 import com.github.fabiomqs.productapi.modules.product.repository.ProductRepository;
+import com.github.fabiomqs.productapi.modules.sales.client.SalesClient;
+import com.github.fabiomqs.productapi.modules.sales.dto.SalesConfirmationDTO;
+import com.github.fabiomqs.productapi.modules.sales.dto.SalesProductResponse;
+import com.github.fabiomqs.productapi.modules.sales.enums.SalesStatus;
+import com.github.fabiomqs.productapi.modules.sales.rabbitmq.SalesConfirmationSender;
 import com.github.fabiomqs.productapi.modules.supplier.service.SupplierService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static com.github.fabiomqs.productapi.config.RequestUtil.getCurrentRequest;
 import static org.springframework.util.ObjectUtils.isEmpty;
 
 @Slf4j
@@ -21,23 +29,27 @@ import static org.springframework.util.ObjectUtils.isEmpty;
 public class ProductService {
 
     private static final Integer ZERO = 0;
-//    private static final String TRANSACTION_ID = "transactionid";
-//    private static final String SERVICE_ID = "serviceid";
+    private static final String TRANSACTION_ID = "transactionid";
+    private static final String SERVICE_ID = "serviceid";
 
 
     private final ProductRepository productRepository;
-
     private final SupplierService supplierService;
-
     private final CategoryService categoryService;
+    private final SalesConfirmationSender salesConfirmationSender;
+    private final SalesClient salesClient;
 
     public ProductService(
             ProductRepository productRepository,
             SupplierService supplierService,
-            CategoryService categoryService) {
+            CategoryService categoryService,
+            SalesConfirmationSender salesConfirmationSender,
+            SalesClient salesClient) {
         this.productRepository = productRepository;
         this.supplierService = supplierService;
         this.categoryService = categoryService;
+        this.salesConfirmationSender = salesConfirmationSender;
+        this.salesClient = salesClient;
     }
 
     public ProductResponse save(ProductRequest request) {
@@ -141,10 +153,10 @@ public class ProductService {
         if (!productRepository.existsById(id)) {
             throw new ValidationException("The product does not exists.");
         }
-        //var sales = getSalesByProductId(id);
-        //if (!isEmpty(sales.getSalesIds())) {
-        //    throw new ValidationException("The product cannot be deleted. There are sales for it.");
-        //}
+        var sales = getSalesByProductId(id);
+        if (!isEmpty(sales.getSalesIds())) {
+            throw new ValidationException("The product cannot be deleted. There are sales for it.");
+        }
         productRepository.deleteById(id);
         return SuccessResponse.create("The product was deleted.");
     }
@@ -155,114 +167,117 @@ public class ProductService {
         }
     }
 
-//    public void updateProductStock(ProductStockDTO product) {
-//        try {
-//            validateStockUpdateData(product);
-//            updateStock(product);
-//        } catch (Exception ex) {
-//            log.error("Error while trying to update stock for message with error: {}", ex.getMessage(), ex);
-//            var rejectedMessage = new SalesConfirmationDTO(product.getSalesId(), SalesStatus.REJECTED, product.getTransactionid());
-//            salesConfirmationSender.sendSalesConfirmationMessage(rejectedMessage);
-//        }
-//    }
-//
-//    private void validateStockUpdateData(ProductStockDTO product) {
-//        if (isEmpty(product)
-//                || isEmpty(product.getSalesId())) {
-//            throw new ValidationException("The product data and the sales ID must be informed.");
-//        }
-//        if (isEmpty(product.getProducts())) {
-//            throw new ValidationException("The sales' products must be informed.");
-//        }
-//        product
-//                .getProducts()
-//                .forEach(salesProduct -> {
-//                    if (isEmpty(salesProduct.getQuantity())
-//                            || isEmpty(salesProduct.getProductId())) {
-//                        throw new ValidationException("The productID and the quantity must be informed.");
-//                    }
-//                });
-//    }
-//
-//    @Transactional
-//    private void updateStock(ProductStockDTO product) {
-//        var productsForUpdate = new ArrayList<Product>();
-//        product
-//                .getProducts()
-//                .forEach(salesProduct -> {
-//                    var existingProduct = findById(salesProduct.getProductId());
-//                    validateQuantityInStock(salesProduct, existingProduct);
-//                    existingProduct.updateStock(salesProduct.getQuantity());
-//                    productsForUpdate.add(existingProduct);
-//                });
-//        if (!isEmpty(productsForUpdate)) {
-//            productRepository.saveAll(productsForUpdate);
-//            var approvedMessage = new SalesConfirmationDTO(product.getSalesId(), SalesStatus.APPROVED, product.getTransactionid());
-//            salesConfirmationSender.sendSalesConfirmationMessage(approvedMessage);
-//        }
-//    }
-//
-//    private void validateQuantityInStock(ProductQuantityDTO salesProduct,
-//                                         Product existingProduct) {
-//        if (salesProduct.getQuantity() > existingProduct.getQuantityAvailable()) {
-//            throw new ValidationException(
-//                    String.format("The product %s is out of stock.", existingProduct.getId()));
-//        }
-//    }
-//
-//    public ProductSalesResponse findProductSales(Integer id) {
-//        var product = findById(id);
-//        var sales = getSalesByProductId(product.getId());
-//        return ProductSalesResponse.of(product, sales.getSalesIds());
-//    }
-//
-//    private SalesProductResponse getSalesByProductId(Integer productId) {
-//        try {
-//            var currentRequest = getCurrentRequest();
-//            var transactionid = currentRequest.getHeader(TRANSACTION_ID);
-//            var serviceid = currentRequest.getAttribute(SERVICE_ID);
-//            log.info("Sending GET request to orders by productId with data {} | [transactionID: {} | serviceID: {}]",
-//                    productId, transactionid, serviceid);
-//            var response = salesClient
-//                    .findSalesByProductId(productId)
-//                    .orElseThrow(() -> new ValidationException("The sales was not found by this product."));
-//            log.info("Recieving response from orders by productId with data {} | [transactionID: {} | serviceID: {}]",
-//                    new ObjectMapper().writeValueAsString(response), transactionid, serviceid);
-//            return response;
-//        } catch (Exception ex) {
-//            throw new ValidationException("The sales could not be found.");
-//        }
-//    }
-//
-//    public SuccessResponse checkProductsStock(ProductCheckStockRequest request) {
-//        try {
-//            var currentRequest = getCurrentRequest();
-//            var transactionid = currentRequest.getHeader(TRANSACTION_ID);
-//            var serviceid = currentRequest.getAttribute(SERVICE_ID);
-//            log.info("Request to POST product stock with data {} | [transactionID: {} | serviceID: {}]",
-//                    new ObjectMapper().writeValueAsString(request), transactionid, serviceid);
-//            if (isEmpty(request) || isEmpty(request.getProducts())) {
-//                throw new ValidationException("The request data and products must be informed.");
-//            }
-//            request
-//                    .getProducts()
-//                    .forEach(this::validateStock);
-//            var response = SuccessResponse.create("The stock is ok!");
-//            log.info("Response to POST product stock with data {} | [transactionID: {} | serviceID: {}]",
-//                    new ObjectMapper().writeValueAsString(response), transactionid, serviceid);
-//            return response;
-//        } catch (Exception ex) {
-//            throw new ValidationException(ex.getMessage());
-//        }
-//    }
-//
-//    private void validateStock(ProductQuantityDTO productQuantity) {
-//        if (isEmpty(productQuantity.getProductId()) || isEmpty(productQuantity.getQuantity())) {
-//            throw new ValidationException("Product ID and quantity must be informed.");
-//        }
-//        var product = findById(productQuantity.getProductId());
-//        if (productQuantity.getQuantity() > product.getQuantityAvailable()) {
-//            throw new ValidationException(String.format("The product %s is out of stock.", product.getId()));
-//        }
-//    }
+    public void updateProductStock(ProductStockDTO product) {
+        try {
+            validateStockUpdateData(product);
+            updateStock(product);
+        } catch (Exception ex) {
+            log.error("Error while trying to update stock for message with error: {}", ex.getMessage(), ex);
+            var rejectedMessage = new SalesConfirmationDTO(
+                    product.getSalesId(),
+                    SalesStatus.REJECTED,
+                    product.getTransactionid());
+            salesConfirmationSender.sendSalesConfirmationMessage(rejectedMessage);
+        }
+    }
+
+    private void validateStockUpdateData(ProductStockDTO product) {
+        if (isEmpty(product)
+                || isEmpty(product.getSalesId())) {
+            throw new ValidationException("The product data and the sales ID must be informed.");
+        }
+        if (isEmpty(product.getProducts())) {
+            throw new ValidationException("The sales' products must be informed.");
+        }
+        product
+                .getProducts()
+                .forEach(salesProduct -> {
+                    if (isEmpty(salesProduct.getQuantity())
+                            || isEmpty(salesProduct.getProductId())) {
+                        throw new ValidationException("The productID and the quantity must be informed.");
+                    }
+                });
+    }
+
+    @Transactional
+    private void updateStock(ProductStockDTO product) {
+        var productsForUpdate = new ArrayList<Product>();
+        product
+                .getProducts()
+                .forEach(salesProduct -> {
+                    var existingProduct = findById(salesProduct.getProductId());
+                    validateQuantityInStock(salesProduct, existingProduct);
+                    existingProduct.updateStock(salesProduct.getQuantity());
+                    productsForUpdate.add(existingProduct);
+                });
+        if (!isEmpty(productsForUpdate)) {
+            productRepository.saveAll(productsForUpdate);
+            var approvedMessage = new SalesConfirmationDTO(product.getSalesId(), SalesStatus.APPROVED, product.getTransactionid());
+            salesConfirmationSender.sendSalesConfirmationMessage(approvedMessage);
+        }
+    }
+
+    private void validateQuantityInStock(ProductQuantityDTO salesProduct,
+                                         Product existingProduct) {
+        if (salesProduct.getQuantity() > existingProduct.getQuantityAvailable()) {
+            throw new ValidationException(
+                    String.format("The product %s is out of stock.", existingProduct.getId()));
+        }
+    }
+
+    public ProductSalesResponse findProductSales(Integer id) {
+        var product = findById(id);
+        var sales = getSalesByProductId(product.getId());
+        return ProductSalesResponse.of(product, sales.getSalesIds());
+    }
+
+    private SalesProductResponse getSalesByProductId(Integer productId) {
+        try {
+            var currentRequest = getCurrentRequest();
+            var transactionid = currentRequest.getHeader(TRANSACTION_ID);
+            var serviceid = currentRequest.getAttribute(SERVICE_ID);
+            log.info("Sending GET request to orders by productId with data {} | [transactionID: {} | serviceID: {}]",
+                    productId, transactionid, serviceid);
+            var response = salesClient
+                    .findSalesByProductId(productId)
+                    .orElseThrow(() -> new ValidationException("The sales was not found by this product."));
+            log.info("Recieving response from orders by productId with data {} | [transactionID: {} | serviceID: {}]",
+                    new ObjectMapper().writeValueAsString(response), transactionid, serviceid);
+            return response;
+        } catch (Exception ex) {
+            throw new ValidationException("The sales could not be found.");
+        }
+    }
+
+    public SuccessResponse checkProductsStock(ProductCheckStockRequest request) {
+        try {
+            var currentRequest = getCurrentRequest();
+            var transactionid = currentRequest.getHeader(TRANSACTION_ID);
+            var serviceid = currentRequest.getAttribute(SERVICE_ID);
+            log.info("Request to POST product stock with data {} | [transactionID: {} | serviceID: {}]",
+                    new ObjectMapper().writeValueAsString(request), transactionid, serviceid);
+            if (isEmpty(request) || isEmpty(request.getProducts())) {
+                throw new ValidationException("The request data and products must be informed.");
+            }
+            request
+                    .getProducts()
+                    .forEach(this::validateStock);
+            var response = SuccessResponse.create("The stock is ok!");
+            log.info("Response to POST product stock with data {} | [transactionID: {} | serviceID: {}]",
+                    new ObjectMapper().writeValueAsString(response), transactionid, serviceid);
+            return response;
+        } catch (Exception ex) {
+            throw new ValidationException(ex.getMessage());
+        }
+    }
+
+    private void validateStock(ProductQuantityDTO productQuantity) {
+        if (isEmpty(productQuantity.getProductId()) || isEmpty(productQuantity.getQuantity())) {
+            throw new ValidationException("Product ID and quantity must be informed.");
+        }
+        var product = findById(productQuantity.getProductId());
+        if (productQuantity.getQuantity() > product.getQuantityAvailable()) {
+            throw new ValidationException(String.format("The product %s is out of stock.", product.getId()));
+        }
+    }
 }
